♻️ refactor: 스타일 분기 처리 및 컴포넌트 이름 수정, 아이콘콘 크기 변경
