+++ components/rollyList/RollyListItem.module.scss
@@ -1,4 +1,4 @@
-.rolly-list {
+.rolly-list-item {
   width: 100%;
   padding: 12px;
   justify-content: center;
@@ -13,6 +13,9 @@
   align-items: center;
   justify-content: space-between;
 }
+.list:nth-child(2) {
+  margin-top: 30px;
+}
 .title {
   font-size: var(--font-size-lg);
   color: var(--color-black);
@@ -23,5 +26,5 @@
 }
 .lock,
 .action {
-  margin-top: 5px;
+  margin-top: 3px;
 }
